Add TierPriority with codegen

+++ pkg/controller/networkpolicy/clusternetworkpolicy.go
@@ -33,7 +33,7 @@ var (
 	}
 
 	// initialTierPriority assigns the lowest priority value for the Tier.
-	initialTierPriority = uint8(255)
+	initialTierPriority = uint32(255)
 )
 
 // addCNP receives ClusterNetworkPolicy ADD events and creates resources
@@ -200,7 +200,7 @@ func (n *NetworkPolicyController) toAntreaPeerForCRD(peers []secv1alpha1.Network
 // getTierPriority returns the Priority associated with the Tier object. If
 // tier name is empty, the policy is associated with the lowest Tier i.e.
 // initial-tier(255).
-func (n *NetworkPolicyController) getTierPriority(tName string) uint8 {
+func (n *NetworkPolicyController) getTierPriority(tName string) uint32 {
 	if tName == "" {
 		// Empty Tier name should move the CNP to the initialTier.
 		return initialTierPriority
@@ -208,7 +208,7 @@ func (n *NetworkPolicyController) getTierPriority(tName string) uint8 {
 	// Retrieve corresponding Tier and return its priority.
 	tObj, _ := n.tierLister.Get(tName)
 	klog.V(2).Infof("Found Tier %s with Priority %d", tName, tObj.Spec.Priority)
-	return tObj.Spec.Priority
+	return uint32(tObj.Spec.Priority)
 }
 
 // createAddressGroupForCRD creates an AddressGroup object corresponding to a
